CI Improvements (#6619)
* dont run regression tests on draft PR

* extend behavior to docsgen as well

* move envoy and guard docs and regression tests

* update readme

* Trigger CI
