+++ .github/workflows/docs-gen.yaml
@@ -14,11 +14,58 @@ on:
     types:
       - completed
 jobs:
+  prepare_env:
+    name: Prepare Environment
+    runs-on: ubuntu-18.04
+    outputs:
+      should-build-docs: ${{ steps.build-docs.outputs.build_value }}
+      should-pass-docs: ${{ steps.build-docs.outputs.pass_value }}
+    steps:
+      - id: is-draft-pr
+        name: Process draft Pull Requests
+        if: ${{ github.event.pull_request.draft }}
+        run: echo "::set-output name=value::$(echo true)"
+      - id: signal-docs-comment
+        name: Process comments on Pull Request to signal Docs
+        if: ${{ github.event.issue.pull_request }}
+        run: echo "::set-output name=value::$(echo ${{ contains(github.event.comment.body, '/sig-docs') }})"
+      - id: skip-docs-comment
+        name: Process comments on Pull Request to skip Docs
+        if: ${{ github.event.issue.pull_request }}
+        run: echo "::set-output name=value::$(echo ${{ contains(github.event.comment.body, '/skip-docs') }})"
+      - id: build-docs
+        name: Determine whether or not to build docs
+        run: |
+          should_build=true
+          should_pass=false
+
+          is_draft_pr=${{ steps.is-draft-pr.outputs.value }}
+          if [[ ! -z $is_draft_pr && $is_draft_pr = true ]] ; then
+            should_build=false
+          fi
+
+          pr_comment_signal=${{ steps.signal-docs-comment.outputs.value }}
+          if [[ ! -z $pr_comment_signal && $pr_comment_signal = true ]] ; then
+            should_build=true
+          fi
+
+          pr_comment_skip=${{ steps.skip-docs-comment.outputs.value }}
+          if [[ ! -z $pr_comment_skip && $pr_comment_skip = true ]] ; then
+            should_pass=true
+          fi
+
+          echo "Should build docs? $should_build"
+          echo "Should auto-pass docs? $should_pass"
+          echo "::set-output name=build_value::$(echo $should_build)"
+          echo "::set-output name=pass_value::$(echo $should_pass)"
   build:
     name: Generate versioned docs site
+    needs: prepare_env
+    if: needs.prepare_env.outputs.should-build-docs == 'true'
     runs-on: ubuntu-18.04
     steps:
     - name: Free disk space
+      if: needs.prepare_env.outputs.should-pass-docs != 'true'
       run: |
         echo "Before clearing disk space:"
         df -h
@@ -36,15 +83,18 @@ jobs:
         echo "After clearing disk space:"
         df -h
     - name: Set up Go
+      if: needs.prepare_env.outputs.should-pass-docs != 'true'
       uses: actions/setup-go@v2
       with:
         go-version: 1.18.2
       id: go
     - name: Setup Hugo
+      if: needs.prepare_env.outputs.should-pass-docs != 'true'
       uses: peaceiris/actions-hugo@v2
       with:
         hugo-version: '0.69.2'
     - name: Check out code into the Go module directory
+      if: needs.prepare_env.outputs.should-pass-docs != 'true'
       uses: actions/checkout@v2
     - name: Detect Community PR
       id: community-pr-check
@@ -54,7 +104,7 @@ jobs:
         echo "Pull Request is from a fork. Setting IS_COMMUNITY_PR to true"
         echo "::set-output name=IS_COMMUNITY_PR::true"
     - name: Generate versioned docs site
-      if: ${{ !steps.community-pr-check.outputs.IS_COMMUNITY_PR }}
+      if: ${{ !steps.community-pr-check.outputs.IS_COMMUNITY_PR && needs.prepare_env.outputs.should-pass-docs != 'true' }}
       run: make -C docs build-site
       env:
         GITHUB_TOKEN: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
@@ -73,7 +123,7 @@ jobs:
         entryPoint: ./docs/ci
     - name: Deploy to Firebase (preview)
       # Generate live preview of docs for PRs to master
-      if: ${{ github.event_name == 'pull_request' && !steps.community-pr-check.outputs.IS_COMMUNITY_PR }}
+      if: ${{ github.event_name == 'pull_request' && !steps.community-pr-check.outputs.IS_COMMUNITY_PR && needs.prepare_env.outputs.should-pass-docs != 'true' }}
       uses: FirebaseExtended/action-hosting-deploy@v0
       with:
         repoToken: ${{ secrets.GITHUB_TOKEN }}
@@ -87,4 +137,4 @@ jobs:
         curl -X POST\
              -H 'Content-type: application/json'\
              --data '{"text":"Gloo Edge has <https://github.com/solo-io/gloo/actions/runs/${{github.run_id}}|failed a docs build> on `master` branch"}'\
-             ${{ env.SLACK_DEBUG_TESTING && secrets.SLACK_INTEGRATION_TESTING_WEBHOOK || secrets.EDGE_TEAM_BOTS_WEBHOOK }}
+             ${{ env.SLACK_DEBUG_TESTING == true && secrets.SLACK_INTEGRATION_TESTING_WEBHOOK || secrets.EDGE_TEAM_BOTS_WEBHOOK }}
