+++ .github/workflows/notify-on-regression-failure-list.json
@@ -10,8 +10,6 @@
     "__jenshu": "U016LC0Q70A",
     "__jmunozro" : "U0214QV6L3Y",
     "kdorosh": "ULE7TSJ12",
-    "__MitchAman": "U01TH9ZJP3N",
-    "mkazin": "U02JFCX4YAE",
     "__nfuden": "U02JCDBKEDT",
     "__Rachael-Graham": "U02BQ7E7W3S",
     "__saiskee": "U01D06A9NVA",
